refactor api of constructing B* and SB* types
previously, the `new` function would return `Option<Self>`, and there
was no panicking constructor. this is inconvenient, and doesn't match
the design of the `BitPiece` trait, where `from_bits` is panicking and
`try_from_bits` returns `Option<Self>`.

change the api of constructing B* and SB* types to be more like the api
of the `BitPiece` trait - the `new` should panic, and the `try_new`
function should return `Option<Self>`.